avoid probes initiating a HTTP sesssion
