+++ pkg/consolegraphql/server/session.go
@@ -9,6 +9,7 @@ package server
 import (
 	"github.com/alexedwards/scs/v2"
 	"github.com/enmasseproject/enmasse/pkg/consolegraphql/server/session"
+	"github.com/prometheus/client_golang/prometheus"
 	"sync/atomic"
 	"time"
 )
@@ -17,25 +18,31 @@ type sessionCountingListener struct {
 	sessionCount      int32
 	sessionPresent    func()
 	noSessionsPresent func()
+	metric            prometheus.Gauge
 }
 
 func (l *sessionCountingListener) Added(string) {
-	if atomic.AddInt32(&l.sessionCount, 1) == 1 {
+	curr := atomic.AddInt32(&l.sessionCount, 1)
+	if curr == 1 {
 		l.sessionPresent()
 	}
+	l.metric.Set(float64(curr))
 }
 
 func (l *sessionCountingListener) Removed(string) {
-	if atomic.AddInt32(&l.sessionCount, -1) == 0 {
+	curr := atomic.AddInt32(&l.sessionCount, -1)
+	if curr == 0 {
 		l.noSessionsPresent()
 	}
+	l.metric.Set(float64(curr))
 }
 
-func CreateSessionManager(sessionLifetime time.Duration, sessionIdleTimeout time.Duration, sessionPresentCallback func(), noSessionsPresentCallback func()) *scs.SessionManager {
+func CreateSessionManager(sessionLifetime time.Duration, sessionIdleTimeout time.Duration, sessionPresentCallback func(), noSessionsPresentCallback func(), metric prometheus.Gauge) *scs.SessionManager {
 	store := session.New()
 	store.RegisterListener(&sessionCountingListener{
 		sessionPresent:    sessionPresentCallback,
 		noSessionsPresent: noSessionsPresentCallback,
+		metric:            metric,
 	})
 
 	sessionManager := scs.New()
